Use JSBeeb's rewind-thumbnail with types instead of copying
